+++ src/main/java/tc/SubPartitionToPointsFlatMapper.java
@@ -10,20 +10,20 @@ import java.util.List;
 import java.util.Map;
 
 public class SubPartitionToPointsFlatMapper implements
-        PairFlatMapFunction<Tuple2<Integer,TCRegion>,
+        PairFlatMapFunction<Tuple2<Long,TCRegion>,
                 String, Tuple2<Integer, TCPoint>>
 {
     @Override
-    public Iterable<Tuple2<String, Tuple2<Integer, TCPoint>>> call(Tuple2<Integer, TCRegion> input) throws Exception {
+    public Iterable<Tuple2<String, Tuple2<Integer, TCPoint>>> call(Tuple2<Long, TCRegion> input) throws Exception {
         return apply(input);
     }
 
-    public Iterable<Tuple2<String, Tuple2<Integer, TCPoint>>> apply(Tuple2<Integer, TCRegion> input) {
+    public Iterable<Tuple2<String, Tuple2<Integer, TCPoint>>> apply(Tuple2<Long, TCRegion> input) {
         List<Tuple2<String, Tuple2<Integer, TCPoint>>> result =
                 new ArrayList();
 
         TCRegion region = input._2();
-        int slotId = input._1();
+        long slotId = input._1();
         int regionId = region.getRegionId();
         String key = String.format("%s,%s", slotId, regionId);
         Map<Integer, TCPoint> pointMap = region.getPoints();
